refactor: extract loadMoods logic to custom useMoods hook.
