+++ screens/HistoryScreen.js
@@ -1,29 +1,15 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
 import { useFocusEffect } from '@react-navigation/native';
 import React, { useCallback, useState } from 'react';
 import { FlatList, Text, View } from 'react-native';
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
+import useMoods from '../hooks/useMoods'
 
 import MoodCard from '../components/MoodCard';
-import { MOOD_LIST_KEY } from '../constants/storage';
 import { sg } from '../styling';
 
 const HistoryScreen = () => {
   const insets = useSafeAreaInsets();
-  const [moodList, setMoodList] = useState([]);
-
-  const loadMoods = async () => {
-    try {
-      const stored = await AsyncStorage.getItem(MOOD_LIST_KEY);
-      if (stored) {
-        const parsed = JSON.parse(stored);
-        const sortedList = [...parsed].sort((a, b) => new Date(b.date) - new Date(a.date));
-        setMoodList(sortedList);
-      }
-    } catch (e) {
-      console.error('Failed to load moods', e);
-    }
-  };
+  const { moodList, loadMoods } = useMoods();
 
   useFocusEffect(
     useCallback(() => {
@@ -40,8 +26,8 @@ const HistoryScreen = () => {
           data={moodList}
           keyExtractor={(item) => item.id.toString()}
           renderItem={({ item }) => <MoodCard item={item} />}
-          style={{ marginTop: 20 }}
-          contentContainerStyle={{ paddingBottom: 100 }}
+          style={sg.mT20}
+          contentContainerStyle={sg.pB100}
           showsVerticalScrollIndicator={false}
         />
       )}
